fix: create notification channels early & make it more granular
