+++ shared/domain/src/commonMain/resources/mobile/mobile.properties
@@ -63,7 +63,8 @@ mobile.action.finish=Finish
 # Notifications
 ################################################################################
 mobile.android.channels.service=Bisq service
-mobile.android.channels.tradeAndUpdates=Trade and updates
+mobile.android.channels.tradeState=Trade updates
+mobile.android.channels.userMessages=User messages
 mobile.action.notifications.openChat=Open Chat
 mobile.action.notifications.openTrade=Open Trade
