+++ shared/presentation/src/commonMain/kotlin/network/bisq/mobile/presentation/notification/AndroidNotificationChannels.kt
@@ -5,5 +5,6 @@ package network.bisq.mobile.presentation.notification
  */
 object AndroidNotificationChannels {
     const val BISQ_SERVICE_CHANNEL_ID = "BISQ_SERVICE_CHANNEL"
-    const val TRADE_AND_UPDATES_CHANNEL_ID = "BISQ_TRADE_AND_UPDATES_CHANNEL"
+    const val TRADE_UPDATES_CHANNEL_ID = "BISQ_TRADE_UPDATES_CHANNEL"
+    const val USER_MESSAGES_CHANNEL_ID = "USER_MESSAGES_CHANNEL"
 }
