+++ shared/presentation/src/commonMain/kotlin/network/bisq/mobile/presentation/ui/uicases/DashboardPresenter.kt
@@ -96,9 +96,5 @@ open class DashboardPresenter(
         launchIO { settingsRepository.setNotificationPermissionState(state) }
     }
 
-    fun doPlatformSpecificSetup() {
-        notificationController.doPlatformSpecificSetup()
-    }
-
     suspend fun hasNotificationPermission(): Boolean = notificationController.hasPermission()
 }
